+++ python/cugraph/structure/graph.py
@@ -275,8 +275,8 @@ class Graph:
                 df = cudf.DataFrame()
                 for c in edgelist_df.columns:
                     if c in ['src', 'dst']:
-                        df[c] = self.edgelist.renumber_map[edgelist_df[c]].\
-                            reset_index(drop=True)
+                        df[c] = self.edgelist.renumber_map.\
+                            iloc[edgelist_df[c]].reset_index(drop=True)
                     else:
                         df[c] = edgelist_df[c]
             return df
@@ -457,10 +457,10 @@ class Graph:
                     + ['second_' + str(i) for i in range(n_cols)]
                 df = unrenumbered_df
             else:
-                df['first'] = self.edgelist.renumber_map[df['first']].\
-                    reset_index(drop=True)
-                df['second'] = self.edgelist.renumber_map[df['second']].\
-                    reset_index(drop=True)
+                df['first'] = self.edgelist.renumber_map.\
+                    iloc[df['first']].reset_index(drop=True)
+                df['second'] = self.edgelist.renumber_map.\
+                    iloc[df['second']].reset_index(drop=True)
         return df
 
     def number_of_vertices(self):
@@ -670,7 +670,7 @@ class Graph:
         df = cudf.DataFrame()
         if vertex_subset is None:
             if self.renumbered is True:
-                df['vertex'] = self.edgelist.renumber_map[vertex_col]
+                df['vertex'] = self.edgelist.renumber_map.iloc[vertex_col]
             else:
                 df['vertex'] = vertex_col
             df['in_degree'] = in_degree_col
@@ -705,7 +705,7 @@ class Graph:
         df = cudf.DataFrame()
         if vertex_subset is None:
             if self.renumbered is True:
-                df['vertex'] = self.edgelist.renumber_map[vertex_col]
+                df['vertex'] = self.edgelist.renumber_map.iloc[vertex_col]
             else:
                 df['vertex'] = vertex_col
             df['degree'] = degree_col
@@ -849,7 +849,7 @@ class Graph:
         df = self.edgelist.edgelist_df
         n = cudf.concat([df['src'], df['dst']]).unique()
         if self.renumbered:
-            return self.edgelist.renumber_map[n]
+            return self.edgelist.renumber_map.iloc[n]
         else:
             return n
 
@@ -865,7 +865,7 @@ class Graph:
         df = self.edgelist.edgelist_df
         neighbors = df[df['src'] == n]['dst'].reset_index(drop=True)
         if self.renumbered:
-            return self.edgelist.renumber_map[neighbors]
+            return self.edgelist.renumber_map.iloc[neighbors]
         else:
             return neighbors